requirements.txt
streamlit
Pillow
opencv-python-headless
tensorflow

